Moves supported events cache into events endpoint
