+++ templates/slack/_files/f/handler/index.js
@@ -13,38 +13,6 @@ const CommandHandler = require('../../slack/handlers/command_handler.js');
 const EventHandler = require('../../slack/handlers/event_handler.js');
 const ActionHandler = require('../../slack/handlers/action_handler.js');
 
-const fs = require('fs');
-const path = require('path');
-
-const findHandledEvents = function() {
-  let eventsDirPath = path.join(__dirname, '..', '..', 'slack', 'events');
-  return fs.readdirSync(eventsDirPath).reduce((handledEvents, filename) => {
-    let eventHandlerDirPath = path.join(eventsDirPath, filename);
-    if (
-      fs.lstatSync(eventHandlerDirPath).isDirectory() &&
-      fs.existsSync(path.join(eventHandlerDirPath, 'handler.js'))
-    ) {
-      handledEvents[filename] = findHandledEventSubtypes(eventHandlerDirPath);
-    }
-    return handledEvents;
-  }, {});
-};
-
-const findHandledEventSubtypes = function(eventHandlerDirPath) {
-  let subtypesDirPath = path.join(eventHandlerDirPath, 'subtypes');
-  if (fs.existsSync(subtypesDirPath) && fs.lstatSync(subtypesDirPath).isDirectory()) {
-    return fs.readdirSync(subtypesDirPath).reduce((handledSubtypes, filename) => {
-      let name = filename.substr(0, filename.lastIndexOf('.js'));
-      handledSubtypes[name] = true;
-      return handledSubtypes;
-    }, {});
-  } else {
-    return {};
-  }
-};
-
-const handledEvents = findHandledEvents();
-
 module.exports = (params, callback) => {
 
   let kwargs = params.kwargs;
@@ -60,13 +28,6 @@ module.exports = (params, callback) => {
   let action = kwargs.action;
 
   if (event) {
-
-    if (
-      !handledEvents[event.type] ||
-      (event.subtype && !handledEvents[event.type][event.subtype])
-    ) {
-      return callback(null, null);
-    }
 
     tokenize(EventHandler, teamId, kwargs, callback);
